pyproject.toml
[tool.poetry]
name = "maximai-client"
version = "0.1.0"
description = ""
authors = ["Julian de Ruiter <julianderuiter@gmail.com>"]
readme = "README.md"
packages = [{include = "maximai_client", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pvcheetah = "^2.0.1"
pvrecorder = "^1.2.2"
pvporcupine = "^3.0.2"
pvorca = "^0.1.4"
wheel = "^0.43.0"
pydub = "^0.25.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
typer = "^0.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
